+++ src/modules/utils/cogs.py
@@ -1,12 +1,13 @@
-from typing import Optional
-
-from modules.utils import logic
-from shared.bot import StrachyBot
+from typing import Optional, List, Tuple
+import datetime
 
 import discord
 from discord.ext import commands
 from discord import app_commands
 
+from shared.bot import StrachyBot
+from shared import console, messages
+from .helpers import parse_changelog
 
 class UtilsCog(commands.Cog):
     def __init__(self, bot: StrachyBot) -> None:
@@ -16,7 +17,32 @@ class UtilsCog(commands.Cog):
     async def info(
         self,
         interaction: discord.Interaction) -> None:
-        await logic.show_info(interaction, self.bot.start_time)
+        console.log_info(f"/info: User {interaction.user.display_name} used the command.")
+
+        try:
+            uptime: datetime.timedelta = discord.utils.utcnow() - self.bot.start_time
+
+            embed: discord.Embed = discord.Embed(
+                color=discord.Color.blue(),
+                title="StrachyBot 🤖",
+                description="Discord bot with fun mini-games like *Wordle* and *Tic-Tac-Toe*."
+            )
+            
+            embed.add_field(name="Ping", value=f"{round(interaction.client.latency * 1000)} ms", inline=True)
+            embed.add_field(name="Uptime", value=f"{str(uptime).split('.')[0]}", inline=True)
+
+            sections: List[Tuple[str, str]] = parse_changelog()
+
+            for section_name, section_content in sections:
+                embed.add_field(name=section_name, value=section_content, inline=False)
+
+            icon: discord.File = discord.File(
+            "./src/modules/utils/info.png", filename="info.png")
+            embed.set_thumbnail(url="attachment://info.png")
+            
+            await interaction.response.send_message(embed=embed, file=icon)
+        except Exception as e:
+            await messages.handle_error(e, interaction)
 
 
     @app_commands.command(name="announcement", description="Make announcements in chat.")
@@ -25,4 +51,19 @@ class UtilsCog(commands.Cog):
         interaction: discord.Interaction,
         title: Optional[str] = "",
         message:  Optional[str] = "") -> None:
-        await logic.announce(interaction, title, message)
+        console.log_info(f"/info: User {interaction.user.display_name} passed title '{title}' and '{message}'.")
+    
+        try:
+            embed: discord.Embed = discord.Embed(color=discord.Color.yellow())
+            if title:
+                embed.title = title
+            if message:
+                embed.description = message
+
+            icon: discord.File = discord.File(
+            "./src/modules/utils/announcement.png", filename="announcement.png")
+            embed.set_thumbnail(url="attachment://announcement.png")
+            
+            await interaction.response.send_message(embed=embed, file=icon)
+        except Exception as e:
+            await messages.handle_error(e, interaction)